Refactor and move Location Subtype DTOs to new packages for organization
