+++ src/main/java/dev/austinbarnes/retailinventorymanagement/location/dto/LocationResponseClientDTO.java
@@ -1,5 +1,7 @@
 package dev.austinbarnes.retailinventorymanagement.location.dto;
 
+import dev.austinbarnes.retailinventorymanagement.location.dto.type.LocationTypeResponseDTO;
+
 import java.util.UUID;
 
 public record LocationResponseClientDTO (
