+++ src/main/java/dev/austinbarnes/retailinventorymanagement/location/dto/hours/LocationHoursResponseAdminDTO.java
@@ -1,4 +1,6 @@
-package dev.austinbarnes.retailinventorymanagement.location.dto;
+package dev.austinbarnes.retailinventorymanagement.location.dto.hours;
+
+import dev.austinbarnes.retailinventorymanagement.location.dto.LocationResponseDTO;
 
 import java.time.LocalTime;
 import java.util.UUID;
