+++ src/main/java/dev/austinbarnes/retailinventorymanagement/location/dto/type/LocationTypeResponseClientDTO.java
@@ -1,4 +1,4 @@
-package dev.austinbarnes.retailinventorymanagement.location.dto;
+package dev.austinbarnes.retailinventorymanagement.location.dto.type;
 
 import java.util.UUID;
